+++ apps/mobile/babel.config.js
@@ -1,7 +1,10 @@
+const nativewind = require('nativewind/babel');
+
 module.exports = function (api) {
   api.cache(true);
+  const { plugins } = nativewind();
   return {
     presets: ['babel-preset-expo'],
-    plugins: ['nativewind/babel'],
+    plugins,
   };
 };